requirements.txt
annotated-types==0.7.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
tqdm==4.66.5
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
